requirements.txt
aiohttp==3.8.1
aiomysql==0.1.1
aiosignal==1.2.0
alembic==1.8.0
anyio==3.6.1
async-timeout==4.0.2
attrs==21.4.0
autopep8==2.0.1
bcrypt==3.2.2
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
configparser==5.3.0
cryptography==37.0.2
databases==0.6.0
decorator==5.1.1
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
et-xmlfile==1.1.0
fastapi==0.78.0
frozenlist==1.3.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
importlib-metadata==6.0.0
importlib-resources==5.10.2
joblib==1.2.0
Mako==1.2.0
MarkupSafe==2.1.1
multidict==6.0.2
mysql-client==0.0.1
mysql-connector-python==8.0.31
mysqlclient==2.1.1
numpy==1.22.4
opencv-python==4.6.0.66
openpyxl==3.0.10
ormar==0.11.2
paho-mqtt==1.6.1
pandas==1.4.2
passlib==1.7.4
Pillow==9.1.1
protobuf==3.20.1
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.8.2
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
qrcode==7.3.1
requests==2.28.1
rsa==4.8
scikit-learn==1.2.2
scipy==1.8.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.19.1
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.12
uvicorn==0.18.2
validators==0.20.0
xlrd==2.0.1
yarl==1.7.2
zipp==3.11.0
